停利設定擴增(StrategyStopWinProfit & StrategyStopWinOffset)，開發中
1.Update:停利設定擴增(StrategyStopWinProfit & StrategyStopWinOffset)，開發中
2.Update:輸入參數增加in修飾，根據MSDN的介紹，能減少記憶體的使用，調整中

+++ GNAy.Capital.Trade/GNAy.Capital.Models/src/AppSettings.cs
@@ -175,11 +175,11 @@ namespace GNAy.Capital.Models
         /// <summary>
         /// 整體停利觸發
         /// </summary>
-        public decimal StrategyStopWinProfit { get; set; }
+        public int StrategyStopWinProfit { get; set; }
         /// <summary>
         /// 整體停利位移
         /// </summary>
-        public decimal StrategyStopWinOffset { get; set; }
+        public int StrategyStopWinOffset { get; set; }
 
         /// <summary>
         /// 送出的委託單，不論是否有收到委回成回
